// src/WebUI/ClientApp/src/app/core/graphql.ts
export type RatingInputType = {
  id: number,
  movieId: number;
  movieName: string;
  comment: string;
  ratingPoint: number;
  userMail: string;
}

export type RatingType = {
  'id': string;
  'movieId': string;
  'movieName': string;
  'comment': string;
  'ratingPoint': string;
  'userMail': string;
}